requirements.txt
streamlit==0.89.0
pandas==1.3.3
numpy==1.21.2
opencv-python==4.5.3
Pillow==8.3.1
face_recognition==1.3.0
dlib==19.22.0
cmake==3.21.3
Flask==2.0.1